+++ packages/ali-sms-plugin/src/sms-event.ts
@@ -4,7 +4,7 @@ export class SmsEvent extends PickerEvent {
     constructor(
         public type: 'verification' | 'content',
         public phone: string,
-        public content: string
+        public content: string,
     ) {
         super()
     }